requirements.txt
torch
torchvision
transformers
ultralytics
opencv-python
numpy
matplotlib
pypcd4
requests
Pillow
